requirements.txt
python==3.8.0
numpy==1.22.2
scikit-image==0.19.0
scikit-learn==1.0.1
importlib_metadata==4.11.1
napari[all]==0.4.9 # fixed
jupyterlab==3.0.13
ipywidgets==7.6.5
albumentations==1.1.0
torch==1.10.2
torchvision==0.11.3
torchinfo==1.6.3
pytorch-lightning==1.4.9 # deliberately choosing < 1.5
neptune-client==0.14.3
neptune-contrib==0.28.1
torchmetrics==0.6.0 # REQUIRED to make dependencies work (changed this from 0.8.1 before) 
